Some minor fixed, added fullscreen mode.

+++ src/main/scala/org/frankversnel/poortjes/util/delta_time.scala
@@ -1,6 +1,6 @@
 package org.frankversnel.poortjes.util
 
-case class DeltaTime(millis: Long) {
+case class DeltaTime(millis: Int) {
 	val seconds: Float = millis.toFloat / 1000f
 }
